requirements.txt
fastapi
jinja2
python-dotenv
pymysql
uvicorn
livereload